data-lake-migration-bigquery-azure/data-lake-migration-bigquery-azure/requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Kaggle API
kaggle>=1.5.0

# BigQuery
google-cloud-bigquery>=3.11.0
google-auth>=2.16.0

# Azure Storage
azure-storage-blob>=12.17.0
azure-storage-file-datalake>=12.12.0
azure-identity>=1.13.0

# Configuration
PyYAML>=6.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.23.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Logging and Monitoring
python-dotenv>=1.0.0

# Data Validation
great-expectations>=0.15.0

# Performance
dask>=2023.5.0
pyarrow>=12.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Model Explainability
shap>=0.42.0
lime>=0.2.0

# MLOps
mlflow>=2.5.0

# Deep Learning (optional)
tensorflow>=2.13.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
